rag-backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
langchain
openai
huggingface_hub
ollama
pydantic
python-multipart
pdfplumber
python-docx
unstructured
sentence-transformers
langchain-huggingface
gunicorn
pandas
requests
pyyaml
google-auth
google-auth-oauthlib
google-api-python-client
langchain-core
langchain-community
langchain_openai
# --- Added from import analysis ---
PyYAML  # for yaml
python-dotenv  # for dotenv
tqdm  # sometimes used for progress bars
# Google API client dependencies (already present):
# google-auth, google-auth-oauthlib, google-api-python-client
# For PDF/docx loaders (already present):
pypdf2
python-docx
# For unstructured email partitioning (already present):
unstructured
openpyxl
langchain-qdrant
# --- Added for Nextcloud integration ---
pyjwt  # Pour la gestion des tokens JWT
webdavclient3  # Pour l'interaction WebDAV avec Nextcloud

# --- Added for expanded file format support ---
# Base unstructured with extensions
unstructured[pdf,doc,ppt,md]  # Extensions for handling different file types
pytesseract  # For OCR capabilities

# Office documents support
python-pptx>=1.0.2  # PowerPoint files
odfpy>=1.4.1  # OpenDocument Text files (ODT)
markdown  # Markdown files

# PDF processing
pdfminer.six  # Enhanced PDF text extraction
pikepdf  # Advanced PDF manipulation
pdf2image  # Convert PDF to images for processing
msal
firebase-admin
minio
# Already included but important for these formats
# python-docx  # Word documents
# openpyxl  # Excel files
# pandas  # CSV and data processing